Refactor: Hide settings menu for non-admin users
This commit modifies the `HomeActivity` to hide the "Settings" menu item in the navigation drawer for users who are not administrators.

The `isAdmin` status is now read from shared preferences. If `isAdmin` is false, the "Settings" menu item is made invisible. The navigation item selected listener is also updated to only allow navigation to settings if the user is an admin.